+++ AbstractInterfaceTask/Models/Rectangular.cs
@@ -6,15 +6,15 @@ namespace AbstractInterfaceTask.Models
 {
     class Rectangular : Figure
     {
-        public int Width { get; set; }
-        public int Length { get; set; }
-        public Rectangular(int width, int length)
+        public double Width { get; set; }
+        public double Length { get; set; }
+        public Rectangular(double width, double length)
         {
             Width = width;
             Length = length;
         }
 
-        public override int CalcArea()
+        public override double CalcArea()
         {
             return Width * Length;
         }
